requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
greenlet==1.1.2
idna==3.3
psycopg2-binary==2.9.3
python-dotenv==0.20.0
requests==2.28.0
SQLAlchemy==1.4.39
urllib3==1.26.9
vk==2.0.2
vk-api==11.9.9
